fix #118423
