+++ src/vs/platform/driver/common/driver.ts
@@ -54,3 +54,12 @@ export interface IWindowDriver {
 	getTerminalBuffer(selector: string): Promise<string[]>;
 	writeInTerminal(selector: string, text: string): Promise<void>;
 }
+
+export interface IDriverOptions {
+	verbose: boolean;
+}
+
+export interface IWindowDriverRegistry {
+	registerWindowDriver(windowId: number): Promise<IDriverOptions>;
+	reloadWindowDriver(windowId: number): Promise<void>;
+}
